% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotGrouped.R
\name{boxplotGrouped}
\alias{boxplotGrouped}
\title{Grouped Boxplots by Factor(s)}
\usage{
boxplotGrouped(
  .data,
  y,
  group.var,
  notch = FALSE,
  y.lab = NULL,
  x.lab = NULL,
  beeswarm = FALSE,
  main = NULL,
  pt.shape = 21,
  pt.size = 2.5
)
}
\arguments{
\item{.data}{Either a \code{data.frame}/\code{tbl_df} object where each column is
a numeric vector containing values for each box, or a \emph{named}
\code{list} object which can be converted to one.}

\item{y}{Character. The response variable for the \code{y} axis.}

\item{group.var}{Character. String of length 1 or 2 representing the
names of the grouping variable(s). Variables must be either factor
or character class vectors.}

\item{notch}{Logical. Should notches be drawn in the boxplots? If FALSE
(default), a standard box plot will be drawn. If TRUE, notches will
be added to the median line of each box plot. See \code{notch} argument of
\code{\link[=geom_boxplot]{geom_boxplot()}}.}

\item{y.lab}{Character. Optional string for the y-axis. Otherwise
one is automatically generated (default).}

\item{x.lab}{Character. Optional string for the x-axis. Otherwise
one is automatically generated (default).}

\item{beeswarm}{Logical. Add points as overlay on top of boxplots?}

\item{main}{Character. Main title for the plot.
See \code{\link[=ggtitle]{ggtitle()}} for \code{ggplot2} style graphics.}

\item{pt.shape}{Numeric or Character. Recognized \code{pch} shapes for the
points. Recall that \emph{only} \code{pch = 21 - 25} are "fill-able". Other point
characters will \emph{not} take on the color from \code{pt.color}.
See \code{\link[=geom_point]{geom_point()}}.}

\item{pt.size}{Numeric. A size for the points. See \code{\link[=geom_point]{geom_point()}}.}
}
\value{
A series of boxplots grouped by \code{group.var}(s).
}
\description{
Generate boxplots of a response variable
by one or two factor variables.
}
\examples{
size <- c("small", "medium", "large")
time <- c("baseline", "6 months", "12 months", "24 months")
df <- data.frame(
  Sample      = sample(size, 40, replace = TRUE),
  TimePoint   = rep(time, each = 10),
  seq.1234.56 = stats::rnorm(40, mean = 25, sd = 3.5)
)

# factor levels determine group order
df$Sample    <- factor(df$Sample, levels = size)
df$TimePoint <- factor(df$TimePoint, levels = time)

# single factor
boxplotGrouped(df, y = "seq.1234.56", group.var = "Sample")

# double factor
boxplotGrouped(df, y = "seq.1234.56", group.var = c("Sample", "TimePoint"))

# with "beeswarm" points
boxplotGrouped(df, y = "seq.1234.56", group.var = "TimePoint", beeswarm = TRUE)
boxplotGrouped(df, y = "seq.1234.56", group.var = c("Sample", "TimePoint"),
               beeswarm = TRUE)
}
\seealso{
\code{\link[=geom_boxplot]{geom_boxplot()}}, \code{\link[=geom_jitter]{geom_jitter()}}

Other boxplots: 
\code{\link{boxplotBeeswarm}()},
\code{\link{boxplotSubarray}()}
}
\author{
Stu Field
}
\concept{boxplots}
