% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotSubarray.R
\name{boxplotSubarray}
\alias{boxplotSubarray}
\title{Plot Boxplots by Subarray (sample)}
\usage{
boxplotSubarray(
  .data,
  color.by = NULL,
  labels = "SampleId",
  y.lim = NULL,
  do.log = TRUE,
  apts = NULL
)
}
\arguments{
\item{.data}{A \code{soma_data} or data frame object created via a call to
\code{\link[=read_adat]{read_adat()}}.}

\item{color.by}{A column name to color the subarrays by. Typically a meta
data field in the adat such as \code{SlideId}.}

\item{labels}{Character. A column name of \code{adat} used to label each box.}

\item{y.lim}{Numeric. Length 2. The upper- and lower-quantiles of the
\emph{total} data used to determine the y-axis limits of the plot.
If \code{NULL}, all points are shown.}

\item{do.log}{Logical. Should the data should be log10-transformed?}

\item{apts}{An optional subset of analytes (as \code{AptNames}) to add on top
of subarray boxplot.}
}
\description{
Plots the distribution of of all analytes stratified
by subarray as a boxplot. In SomaScan (\code{soma_adat}) data format,
a subarray is typically a row or sample in the data.
}
\examples{
data <- SomaDataIO::example_data

# Randomly select a small subset of samples
s_rn <- withr::with_seed(101, sample(rownames(data), 30L))
data <- data[s_rn, ]
boxplotSubarray(data)

# Color by `SampleType` variable
boxplotSubarray(data, color.by = "SampleType")

# Find the feature names of the corresponding hyb controls
hybs <- SomaDataIO::getAnalyteInfo(data) |>
  dplyr::filter(grepl("^Hybridization", Type)) |>
  dplyr::pull(AptName)

# Pass hyb controls to `apts` arg; plots the points in boxes/whiskers
boxplotSubarray(data, color.by = "SampleType", apts = hybs)

# Zoom to (20, 80) quantiles
boxplotSubarray(data, color.by = "SampleType", y.lim = c(0.2, 0.8))
}
\seealso{
\code{\link[=geom_boxplot]{geom_boxplot()}}

Other boxplots: 
\code{\link{boxplotBeeswarm}()},
\code{\link{boxplotGrouped}()}
}
\author{
Stu Field
}
\concept{boxplots}
