% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFoldchangeMatrix.R
\name{calcFoldchangeMatrix}
\alias{calcFoldchangeMatrix}
\alias{print.foldchange_matrix}
\title{Calculate Fold-Change Matrix}
\usage{
calcFoldchangeMatrix(
  data,
  anno_tbl = NULL,
  apt.order = c(NA, "dilution", "cluster"),
  sample.order = NULL,
  max = 3,
  threshold = NULL
)

\method{print}{foldchange_matrix}(x, ...)
}
\arguments{
\item{data}{A \code{soma_adat} object containing RFU feature data.}

\item{anno_tbl}{An annotations table produced via \code{\link[=getAnalyteInfo]{getAnalyteInfo()}}.
Used to calculate analyte dilutions for the matrix column ordering.
If \code{NULL}, a table is generated internally from \code{data} (if possible), and
the analytes are plotted in dilution order.}

\item{apt.order}{Character. How should the columns (features) be ordered. By
dilution mix, clustered, or as is in the \code{data} object?}

\item{sample.order}{Either a character string indicating the column name
with entries to be used to order the data frame rows, or a numeric vector
representing the order of the data frame rows. The
default (\code{NULL}) leaves the row ordering as it is in \code{data}.}

\item{max}{Max value for the range to evaluate. Defaults to 3 which is
typically good for most RFU data.}

\item{threshold}{A threshold for the fold-change values. Values less than
the threshold are set to zero. Threshold must be less than \code{"max"}.}

\item{x}{An object of class \code{"foldchange_matrix"}.}

\item{...}{Arguments for S3 print methods.}
}
\value{
A list of class \code{c("foldchange_matrix", "Map")} containing:
\item{matrix}{A matrix containing the protein fold-change values of each
sample.}
\item{x.lab}{A character string containing the plot x-axis label.}
\item{title}{A character string containing the plot title.}
\item{sample.order}{A numeric vector representing the order of the data
frame rows.}
\item{legend.sub}{A character string containing the plot legend subtitle.}
}
\description{
Calculate the column-wise fold-change matrix of an ADAT.
The clustering method "OLO" (Optimal-leaf-ordering)
produces an optimal leaf ordering with respect to the
minimizing the sum of the distances along the (Hamiltonian)
path connecting the leaves in the given order. The time
complexity of the algorithm is O(n^3). Note that
non-finite distance values are not allowed.
}
\details{
For the S3 plotting method, see \code{\link[=plot.Map]{plot.Map()}}.
}
\section{Functions}{
\itemize{
\item \code{print(foldchange_matrix)}: There is a S3 print method for \code{"foldchange_matrix"}.

}}
\examples{
sample.adat <- SomaDataIO::example_data
fc <- calcFoldchangeMatrix(sample.adat)
class(fc)

# S3 print method
fc

# Sample Order
# specified by user
fc <- calcFoldchangeMatrix(sample.adat, sample.order = 192:1)  # by row indices
fc$sample.order

# specified by field
fc <- calcFoldchangeMatrix(sample.adat, sample.order = "Sex")
fc$sample.order

# specified by 2 fields in ADAT
fc <- calcFoldchangeMatrix(sample.adat,
                           sample.order = c("Sex", "TimePoint"))
fc$sample.order
}
\seealso{
Other Calc Map: 
\code{\link{calcMissingnessMap}()},
\code{\link{calcOutlierMap}()},
\code{\link{getFlaggedIds}()},
\code{\link{plot.Map}()}
}
\author{
Stu Field
}
\concept{Calc Map}
