% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMissingnessMap.R
\name{calcMissingnessMap}
\alias{calcMissingnessMap}
\alias{print.missingness_map}
\title{Calculate Missingness of Clinical Meta Data}
\usage{
calcMissingnessMap(data, include.pattern = ".", exclude.pattern = NULL)

\method{print}{missingness_map}(x, ...)
}
\arguments{
\item{data}{A data frame (or \code{soma_adat}) of \emph{only} the meta data columns
(i.e. no analytes).}

\item{include.pattern}{Character (optional). A regular expression string
used in a \code{\link[=grep]{grep()}} call to include matching column names.
Defaults to \emph{include} all column names in the meta data (".").}

\item{exclude.pattern}{Character (optional). A regular expression string
used in a \code{\link[=grep]{grep()}} call to \emph{exclude} matching column names.}

\item{x}{An object of class \code{"missingness_map"}.}

\item{...}{Arguments for S3 print methods.}
}
\value{
A list of class \code{c("missingness_map", "Map")} containing:
\item{matrix}{A boolean matrix of \code{TRUE/FALSE} whether each sample is in
missingness according the the stated criteria.}
\item{names}{A character vector containing the names of the meta data
columns.}
\item{rows.by.freq}{A logical indicating if the samples are ordered
by missingness frequency. Currently always FALSE.}
\item{legend.sub}{A character string containing the plot legend subtitle.}
\item{title}{A character string containing the plot title.}
\item{x.lab}{A character string containing the plot x-axis label.}
}
\description{
Calculate the "missingness" (\code{NAs}) of clinical meta data of an ADAT.
For the S3 plotting method, see \code{\link[=plot.Map]{plot.Map()}}.
}
\section{Functions}{
\itemize{
\item \code{print(missingness_map)}: There is a S3 print method for \code{"missingness_map"}.

}}
\examples{
sample.adat <- SomaDataIO::example_data
meta <- sample.adat[, SomaDataIO::getMeta(sample.adat)]

# random assign NAs
cols <- rep(1:ncol(meta), each = 3)
rows <- as.integer(replicate(ncol(meta), sample(1:nrow(meta), 3)))
meta[cbind(rows, cols)] <- NA

mm <- calcMissingnessMap(meta)
class(mm)

# S3 print method
mm
}
\seealso{
Other Calc Map: 
\code{\link{calcFoldchangeMatrix}()},
\code{\link{getFlaggedIds}()},
\code{\link{plot.Map}()}
}
\author{
Stu Field
}
\concept{Calc Map}
