% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure.R
\name{figure}
\alias{figure}
\alias{close_figure}
\title{Saves a Figure (Plot) to File}
\usage{
figure(file, height = 480, width = 480, scale = 1, ...)

close_figure(file)
}
\arguments{
\item{file}{Character. The path of the output file passed to \code{\link[=png]{png()}},
\code{\link[=pdf]{pdf()}}, or \code{\link[=jpeg]{jpeg()}}. Plot type determined by file extension.}

\item{height}{Double. The height of the plot in pixels.}

\item{width}{Double. The width of the plot in pixels.}

\item{scale}{A re-scaling of the output to resize window better.}

\item{...}{Additional arguments passed to \code{\link[=png]{png()}}, \code{\link[=pdf]{pdf()}}, or \code{\link[=jpeg]{jpeg()}}.}
}
\value{
The \code{file} argument, invisibly.
}
\description{
A wrapper for \code{\link[=png]{png()}}, \code{\link[=pdf]{pdf()}}, or \code{\link[=jpeg]{jpeg()}} to save plots to
disk. If a file path is passed to \code{\link[=figure]{figure()}}, it
opens a plotting device based on the file extension,
passing the same file name to \code{\link[=close_figure]{close_figure()}}.
If \code{file = NULL}, output is directed to the default plotting device.
}
\details{
The \code{\link[=figure]{figure()}} and \code{\link[=close_figure]{close_figure()}} functions
are most useful when used inside of another function that creates a plot.
By adding a \verb{file =} pass-through argument to a function that creates a plot,
the user can toggle between plotting to file or to a graphics device.
Supported plotting devices:
\itemize{
\item \code{\link[=png]{png()}}
\item \code{\link[=pdf]{pdf()}}
\item \code{\link[=jpeg]{jpeg()}}
\item \code{\link[=postscript]{postscript()}} (\verb{*.eps})
}
}
\section{Functions}{
\itemize{
\item \code{close_figure()}: Closes the currently active plotting device with a
\code{\link[=dev.off]{dev.off()}} call if a file name is passed. If
\code{file = NULL}, nothing happens. This function is typically used in
conjunction with \code{\link[=figure]{figure()}} inside the enclosing function. See example.

}}
\note{
The \code{fontsize} of the plots are constant. If you would like to
increase the font size relative to the plot, you can decrease the plot size.
Alternatively, you can pass \code{pointsize} as an additional argument.
}
\examples{
# Create enclosing plotting function
createPlot <- function(file = NULL) {
  figure(file = file)
  on.exit(close_figure(file = file))
  plot_data <- withr::with_seed(1, matrix(rnorm(30), ncol = 2))
  plot(as.data.frame(plot_data), col = unlist(soma_colors), pch = 19, cex = 2)
}

# default; no file saved
createPlot()

if ( interactive() ) {
  # Save as *.pdf
  createPlot("example.pdf")

  # Save as *.png
  createPlot("example.png")
}
}
\seealso{
\code{\link[=png]{png()}}, \code{\link[=pdf]{pdf()}}, \code{\link[=dev.off]{dev.off()}}

Other base R: 
\code{\link{plotPolygon}()}
}
\author{
Stu Field
}
\concept{base R}
