% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFlaggedIds.R
\name{getFlaggedIds}
\alias{getFlaggedIds}
\title{Get Flagged Ids From MAD Outlier or Missingness Maps}
\usage{
getFlaggedIds(x, flags = 0.05, data = NULL, include = NULL)
}
\arguments{
\item{x}{An object of class: \code{missingness_map} - from \code{\link[=calcMissingnessMap]{calcMissingnessMap()}}}

\item{flags}{Numeric in \verb{[0, 1]}.
For \code{"missingness_map"} class, the percent of "missingness" (\code{NAs}) of the
clinical meta data. If \code{NULL} (default), \code{0.05} (5\%) is selected.}

\item{data}{Optional. The data originally used to create the map \code{x}. If
omitted, a single column data frame is returned.}

\item{include}{Optional. Character vector of column name(s) in \code{data} to
include in the resulting data frame. Ignored if \code{data = NULL}.}
}
\value{
A \code{data.frame} of the indices (\code{idx}) of flagged samples, along
with any additional variables as specified by \code{include}.
}
\description{
Return the IDs of flagged samples for objects of the
\code{missingness_map} class. Samples are flagged based on the percent of
"missingness" (\code{NAs}) of clinical meta data of an ADAT.
}
\examples{
# flagged missingness
sample.adat <- SomaDataIO::example_data
nc   <- SomaDataIO::getMeta(sample.adat, n = TRUE)
meta <- sample.adat[, SomaDataIO::getMeta(sample.adat)]

# random assign NAs to df b/c currently none missing
cols <- rep(1:nc, each = 3L)
rows <- withr::with_seed(1, as.integer(replicate(nc, sample(1:nrow(meta), 3))))
meta[cbind(rows, cols)] <- NA

mm <- calcMissingnessMap(meta)
getFlaggedIds(mm, meta, flags = 0.2)
}
\seealso{
Other Calc Map: 
\code{\link{calcFoldchangeMatrix}()},
\code{\link{calcMissingnessMap}()},
\code{\link{plot.Map}()}
}
\author{
Caleb Scheidel
}
\concept{Calc Map}
