% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{objects}
\alias{objects}
\alias{soma_colors2}
\alias{soma_colors}
\alias{soma_colors_risk}
\alias{soma_colors_greys}
\alias{soma_blue}
\alias{soma_purple}
\alias{col_string}
\alias{gg}
\title{\pkg{SomaPlotr} Colors and Graphics}
\source{
SomaLogic Operating Co., Inc.
}
\usage{
soma_colors2

soma_colors

soma_colors_risk

soma_colors_greys

soma_blue

soma_purple

col_string

gg
}
\description{
This package contains numerous color palettes, themes, and base
\pkg{ggplot2} graphics objects that are used to coordinate and display
a consistent visual graphics presentation.
In most cases the color palettes have been determined by the SomaLogic
Commercial and Marketing teams.
The \pkg{ggplot2} objects are intended to be used for palette and
theme testing. See the description of each object below.
}
\section{soma_colors2}{

The new official \code{"Soma-Colors"} after a 2020 re-branding.
Includes a new \code{"Soma-Blue"} plus other blues, a greens, yellow, pink, and purple:
\tabular{ll}{
Color       \tab Hex-code \cr
\code{blue}      \tab \verb{#4067E2}\cr
\code{teal}      \tab \verb{#1FA8BC}\cr
\code{pink}      \tab \verb{#BC10E0}\cr
\code{yellow}    \tab \verb{#FF8B04}\cr
\code{green}     \tab \verb{#14753D}\cr
\code{turquoise} \tab \verb{#02444F}\cr
\code{lightblue} \tab \verb{#0051D2}\cr
\code{purple}    \tab \verb{#170BA5}
}
Note the following common color-to-group mappings:
\tabular{lcl}{
Controls/Non-Events/Placebo \tab \verb{-->} \tab blue \cr
Cases/Events/Intervention   \tab \verb{-->} \tab teal \cr
Female                      \tab \verb{-->} \tab yellow \cr
Male                        \tab \verb{-->} \tab green
}
}

\section{soma_colors}{

The former official \code{"Soma-Colors"},
includes \code{"Soma-Purple"} plus a green, blue, and two shades of grey:
\tabular{ll}{
Color        \tab Hex-code \cr
\code{purple}     \tab \verb{#24135F}\cr
\code{lightgreen} \tab \verb{#00A499}\cr
\code{lightgrey}  \tab \verb{#707372}\cr
\code{magenta}    \tab \verb{#840B55}\cr
\code{lightblue}  \tab \verb{#006BA6}\cr
\code{yellow}     \tab \verb{#D69A2D}\cr
\code{darkgreen}  \tab \verb{#007A53}\cr
\code{darkblue}   \tab \verb{#1B365D}\cr
\code{darkgrey}   \tab \verb{#54585A}\cr
\code{blue}       \tab \verb{#004C97}
}
}

\section{soma_colors_risk}{

The official \code{"Soma-Risk Colors"} for risk designations.
Includes red, orange, yellow, green, blue:
\tabular{ll}{
Color    \tab Hex-code \cr
\code{red}    \tab \verb{#B33D26}\cr
\code{orange} \tab \verb{#D57800}\cr
\code{yellow} \tab \verb{#D9C756}\cr
\code{green}  \tab \verb{#5CAA7F}\cr
\code{blue}   \tab \verb{#004C97}
}
}

\section{soma_colors_greys}{

The official neutral palette of grey-scale colors:
\tabular{ll}{
Color        \tab Hex-code \cr
\code{black}      \tab \verb{#000000}\cr
\code{darkgrey2}  \tab \verb{#333333}\cr
\code{darkgrey1}  \tab \verb{#444444}\cr
\code{grey}       \tab \verb{#9DABB2}\cr
\code{lightgrey2} \tab \verb{#C4CDD1}\cr
\code{lightgrey1} \tab \verb{#EEF1F2}
}
}

\section{soma_blue}{
 The official \code{"Soma-Blue"} color.
}

\section{soma_purple}{

The former official \code{"Soma-Purple"} color (retired in 2020).
Also the final entry in \link{col_string}.
}

\section{col_string}{

A vector of colors used (sometimes internally) in various plotting routines:
\preformatted{
  [1] "dodgerblue" [2] "red"     [3] "darkgreen" [4] "darkorchid4"
  [5] "cyan"       [6] "orange"  [7] "black"     [8] "grey"
  [9] "#990066"   [10] "green"  [11] "#5A556E"
}
}

\section{gg}{
 A list of 3 sample \code{ggplot} figures based on the \code{iris} data set for
use in testing various themes, color palettes, and general plotting:
\itemize{
\item A scatter plot
\item A bar plot
\item A boxplot
}
}

\examples{
barplot(rep_len(1, length(soma_colors2)), axes = FALSE,
        col = unlist(soma_colors2),
        names = names(soma_colors2), las = 2)

barplot(rep_len(1, length(soma_colors)), axes = FALSE,
        col = unlist(soma_colors),
        names = names(soma_colors), las = 2)

barplot(rep_len(1, length(soma_colors_risk)), axes = FALSE,
        col = unlist(soma_colors_risk),
        names = names(soma_colors_risk), las = 2)

barplot(rep_len(1, length(soma_colors_greys)), axes = FALSE,
        col = unlist(soma_colors_greys),
        names = names(soma_colors_greys), las = 2)

barplot(c(1, 1), col = c(soma_blue, ggplot2::alpha(soma_blue, 0.5)), axes = FALSE,
        names = paste("soma_blue", c("(alpha = 1)", "(alpha = 0.5)")))

barplot(c(1, 1), col = c(soma_purple, ggplot2::alpha(soma_purple, 0.5)), axes = FALSE,
        names = paste("soma_purple", c("(alpha = 1)", "(alpha = 0.5)")))

barplot(rep_len(1, length(col_string)), axes = FALSE,
        col = col_string, names = col_string, las = 2)

gg$point
gg$bar
gg$box
gg$point + ggplot2::theme_dark()
gg$bar + ggplot2::theme_classic()
}
\keyword{datasets}
