% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCDF.R, R/plotCDFbyGroup.R, R/plotCDFlist.R
\name{plotCDF}
\alias{plotCDF}
\alias{plotCDFbyGroup}
\alias{plotCDFlist}
\title{Plots an Empirical CDF}
\usage{
plotCDF(
  x,
  col = soma_colors$purple,
  x.lab = "values",
  y.lab = bquote(~P(X < x)),
  main = "Empirical Cumulative Distribution Function",
  lty = "solid",
  add.gauss = FALSE
)

plotCDFbyGroup(
  data,
  apt,
  group.var,
  cols,
  xlim = NULL,
  x.lab = bquote(italic(log)[10] ~ (RFU)),
  main = apt,
  lty = "solid",
  ablines = NULL,
  include.counts = FALSE
)

plotCDFlist(.data, label = "Group", main = "CDF by Group", ...)
}
\arguments{
\item{x}{A numeric vector.}

\item{col}{Character. String for the color of the line.}

\item{x.lab}{Character. Optional string for the x-axis. Otherwise
one is automatically generated (default).}

\item{y.lab}{Character. Optional string for the y-axis. Otherwise
one is automatically generated (default).}

\item{main}{Character. Main title for the plot.
See \code{\link[=ggtitle]{ggtitle()}} for \code{ggplot2} style graphics.}

\item{lty}{Character. Passed to \code{\link[=geom_vline]{geom_vline()}}. See \code{\link[=ggtitle]{ggtitle()}}.}

\item{add.gauss}{Logical. Should a Gaussian fit of the
data be plotted with the PDF?}

\item{data}{A \code{soma_adat} or \code{data.frame} object containing RFU data.}

\item{apt}{Character. The name of a column in \code{data} to use in generating
CDFs. Typically an aptamer measurement.}

\item{group.var}{An unquoted column name from \code{data} containing group labels.}

\item{cols}{Character. A \emph{vector} of colors for the groups/boxes.
For \code{\link[=plotDoubleHist]{plotDoubleHist()}}, must be \code{length = 2}.}

\item{xlim}{Numeric. Limits for the x-axis.
See \code{\link[=coord_cartesian]{coord_cartesian()}}.}

\item{ablines}{Numeric. A vector of x-axis position(s) for vertical
lines to be added to the CDF or PDF.}

\item{include.counts}{Logical. Should class counts be added to the plot
legend?}

\item{.data}{A \emph{named} list of numeric vectors (or a data.frame)
containing the data to plot.}

\item{label}{Character. A label for the grouping
variable, i.e. what the entries of the list represent.}

\item{...}{Additional arguments passed to either
\code{\link[=plotCDFbyGroup]{plotCDFbyGroup()}} or \code{\link[=plotPDFbyGroup]{plotPDFbyGroup()}}, primarily one of:
\itemize{
\item \code{x.lab}
\item \code{lty}
\item \code{cols}
\item \code{xlim}
\item \code{fill}
\item \code{ablines}
}}
}
\description{
\code{\link[=plotCDF]{plotCDF()}} creates a plot of the \emph{empirical} cumulative
distribution function for a numeric vector of continuous data. It is
similar to \code{\link[=ecdf]{ecdf()}} with some visual modifications.

\code{\link[=plotCDFbyGroup]{plotCDFbyGroup()}} creates a plot with a series of cumulative distribution
function (CDF) plots stratified/split by a grouping
variable of (usually) meta data, e.g. \code{Response} grouping.

\code{\link[=plotCDFlist]{plotCDFlist()}} creates a series of cumulative distribution
function (CDF) plots from a numeric list object.
}
\examples{
# `plotCDF()`
x <- rnorm(100, mean = 5)
plotCDF(x)
plotCDF(x, col = "darkred")
plotCDF(x, col = "dodgerblue", add.gauss = TRUE)

# `plotCDFbyGroup()`
data  <- SomaDataIO::example_data |> dplyr::filter(SampleType == "Sample")
anno  <- SomaDataIO::getTargetNames(SomaDataIO::getAnalyteInfo(data))
fsh   <- "seq.3032.11"
title <- anno[[fsh]]
plotCDFbyGroup(log10(data), apt = fsh, group.var = Sex, main = title)
lines <- split(log10(data[[fsh]]), data$Sex) |>
  vapply(median, double(1))
plotCDFbyGroup(log10(data), apt = fsh, group.var = Sex,
               ablines = lines, main = title)
plotCDFbyGroup(log10(data), apt = fsh, group.var = Sex,
               include.counts = TRUE, main = title)

# `plotCDFlist()`
x <- withr::with_seed(101,
  list(
    Group1 = rnorm(100),
    Group2 = rnorm(100, sd = 0.5),
    Group3 = rnorm(50, mean = 3, sd = 2)
  )
)
# warning: RFU values should all be positive!
plotCDFlist(x)
plotCDFlist(x, label = "SplitBy")
plotCDFlist(x, x.lab = "My x-axis", main = "Variable `x` CDF")
medians <- vapply(x, median, 0.0)
plotCDFlist(x, ablines = medians, main = "Variable `x` CDF")
}
\seealso{
\code{\link[=stat_ecdf]{stat_ecdf()}}, \code{\link[=pnorm]{pnorm()}}

Other cdf-pdf-plots: 
\code{\link{plotPDF}()}
}
\author{
Stu Field
}
\concept{cdf-pdf-plots}
