% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotConcord.R
\name{plotConcord}
\alias{plotConcord}
\alias{plotConcord.numeric}
\alias{plotConcord.data.frame}
\title{Plot Concordance}
\usage{
plotConcord(x, ...)

\method{plotConcord}{numeric}(x, y, ...)

\method{plotConcord}{data.frame}(
  x,
  x.lab = "x",
  y.lab = "y",
  main = NULL,
  pt.col = soma_purple,
  pt.size = 2,
  add.rug = TRUE,
  identify = FALSE,
  all.labels = seq_len(nrow(x)),
  spread = 1,
  label.size = 3,
  cc.type = c("ccc", "pearson", "spearman", "kendall"),
  ...
)
}
\arguments{
\item{x}{Numeric. First vector comparison (x-axis). Alternatively, \code{x} can
be a \strong{two column} \verb{data frame} containing vector data to compare.}

\item{...}{For extensibility to S3 methods.}

\item{y}{Numeric. Second vector comparison (y-axis).
Omitted if \code{x} is a \code{data.frame}.}

\item{x.lab}{Character. Optional string for the x-axis. Otherwise
one is automatically generated (default).}

\item{y.lab}{Character. Optional string for the y-axis. Otherwise
one is automatically generated (default).}

\item{main}{Character. Main title for the plot.
See \code{\link[=ggtitle]{ggtitle()}} for \code{ggplot2} style graphics.}

\item{pt.col}{Character. The color to be used for the points.}

\item{pt.size}{Numeric. The point size.}

\item{add.rug}{Logical. If \code{TRUE}, axis includes ticks for each point.}

\item{identify}{Logical. If \code{TRUE}, points are auto-identified.}

\item{all.labels}{Character. The labels for \strong{each} point if
\code{identify = TRUE}. The points that are actually labeled is determined
by the \verb{spread =} argument. The default simply numbers the points.}

\item{spread}{Numeric. The width from the unit line to start
labeling points. The default (\code{spread = 1}) is reasonable for
log10-transformed RFU data.}

\item{label.size}{Numeric. The size for the point labels.}

\item{cc.type}{Character. The correlation coefficient estimator to use.
Must be one of \code{c("ccc", "pearson", "spearman", "kendall")}.}
}
\value{
A concordance plot with the squared correlation coefficient
and related statistics as a sub-title, a \strong{robust} linear regression
fit line, the unit line, and optionally labeled points.
}
\description{
A concordance plot of 2 numeric vectors, passed either as a
2-column \code{data.frame}/\code{tibble} object \emph{or} separately as \code{x} and \code{y}.

For \code{cc.type = "ccc"}, Lin's Concordance Correlation Coefficient as
is used to estimate the correlation coefficient.
For all other allowed values of \code{cc.type},
\code{\link[stats:cor.test]{stats::cor.test()}} is used.
}
\examples{
x <- withr::with_seed(1234L, rnorm(500, 400, 30))

# bias
plotConcord(x, x + 20)

# scale
plotConcord(x, x * 1.2)

# random
y <- x + withr::with_seed(2345L, rnorm(500, sd = sd(x) / 2))
plotConcord(x, y - (y * 0.02))

# using Spearman's Correlation Coefficient
plotConcord(x, y - (y * 0.02), cc.type = "spearman")

# using Pearson's Correlation Coefficient
plotConcord(x, y - (y * 0.02), cc.type = "pearson")

# using Kendall's Correlation Coefficient
plotConcord(x, y - (y * 0.02), cc.type = "kendall")

# Label points by number (default)
plotConcord(x, y - (y * 0.02), identify = TRUE, spread = 25)

# S3 data frame method
df <- data.frame(x = x, y = y - (y * 0.02))
plotConcord(df, identify = TRUE, spread = 25)
}
\references{
Lin, Lawrence I-Kuei. 1989. A Concordance Correlation
Coefficient to Evaluate Reproducibility. \strong{Biometrics}. 45:255-268.
}
\seealso{
\code{\link[=cor]{cor()}}, \code{\link[=rlm]{rlm()}}, \code{\link[=cor.test]{cor.test()}}
}
\author{
Stu Field
}
