% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDoubleHist.R
\name{plotDoubleHist}
\alias{plotDoubleHist}
\title{Plot a Double-Histogram}
\usage{
plotDoubleHist(
  data,
  cols,
  label = "Group",
  x.lab = "value",
  binwidth = 0.05,
  main = NULL,
  outline = FALSE
)
}
\arguments{
\item{data}{A 2-column \code{data.frame} of numeric
data to be plotted as a paired density histogram.
The column names of the \code{data.frame}
are used as the group names for the plot.}

\item{cols}{Character. A \emph{vector} of colors for the groups/boxes.
For \code{\link[=plotDoubleHist]{plotDoubleHist()}}, must be \code{length = 2}.}

\item{label}{Character. A label for the grouping variable.}

\item{x.lab}{Character. Optional string for the x-axis. Otherwise
one is automatically generated (default).}

\item{binwidth}{Numeric. Set the bin width for the histogram bars.
See \code{\link[=geom_histogram]{geom_histogram()}}.}

\item{main}{Character. Main title for the plot.
See \code{\link[=ggtitle]{ggtitle()}} for \code{ggplot2} style graphics.}

\item{outline}{Logical. Black and white outlines of the histograms.}
}
\description{
Plots \emph{two} density histograms side-by-side.
}
\examples{
dat <- withr::with_seed(123, data.frame(a = rnorm(1000, 2, 0.3),
                                        b = rnorm(1000, 3, 0.3)))
plotDoubleHist(dat)
plotDoubleHist(dat, label = "Grouped By", main = "Two Distributions")
plotDoubleHist(dat, label = "Grouped By", main = "Black & White", outline = TRUE)
plotDoubleHist(dat, binwidth = 0.01)
plotDoubleHist(dat, binwidth = 0.01, outline = TRUE)
}
\author{
Stu Field
}
