% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotLongitudinal.R
\name{plotLongitudinal}
\alias{plotLongitudinal}
\title{Plot Longitudinal Samples by Subject}
\usage{
plotLongitudinal(
  data,
  y,
  time,
  id,
  color.by = NULL,
  summary.line = stats::median,
  size = 2.5,
  main = NULL,
  y.lab = bquote(italic(log)[10] ~ (RFU)),
  x.lab = NULL,
  add.box = FALSE
)
}
\arguments{
\item{data}{A data frame containing RFU values to plot, typically
a \code{soma_adat} class object.}

\item{y}{A quoted or unquoted variable name in \code{data} for the column
containing the y-axis (i.e. typically the "Response" variable).}

\item{time}{A quoted or unquoted variable name in \code{data} for the column
containing the x-axis (i.e. typically the "time" variable).}

\item{id}{A quoted or unquoted variable name in \code{data} for the column
containing the subject IDs (i.e. how the samples should be connected).}

\item{color.by}{A quoted or unquoted variable name in \code{data} for the column
indicating how to color/group the lines.}

\item{summary.line}{A function describing how to summarize the lines.
Typically \code{\link[=mean]{mean()}} or \code{\link[=median]{median()}}. Set to \code{NULL} to suppress summary lines.}

\item{size}{Numeric. The size for the points on the subject lines.}

\item{main}{Character. Main title for the plot.
See \code{\link[=ggtitle]{ggtitle()}} for \code{ggplot2} style graphics.}

\item{y.lab}{Character. Optional string to set the y-axis label.
Defaults to \code{log10(RFU)} if a \code{SeqId} name is detected for \code{y}.}

\item{x.lab}{Character. Optional string for the x-axis. Otherwise
one is automatically generated (default).}

\item{add.box}{Logical. Should boxplots be drawn for each time point?
\strong{Note}: this groups the subjects together by time point and looks only
at the differences across time points, thus \code{time} \emph{must} be a factor.}
}
\value{
A longitudinal plot of samples by subject.
}
\description{
A samples plot by grouping the subjects across time in a
longitudinal study. Can be as few as two longitudinal samples
(e.g. paired samples). See \emph{examples} for various plotting options.
In order to generate a longitudinal plot by sample, 4 variables
(columns of the data frame \code{data}) \emph{must} be provided:
\itemize{
\item \code{y}: Typically the field containing the y-axis.
\item \code{time}: The field containing the time-series information, the x-axis.
\item \code{id}: The field containing \code{subject} identifiers, i.e. how the points
in each line should be connected.
\item \code{color.by}: A field describing how groups should be split/colored. The
only argument with a default value.
}
}
\examples{
df <- withr::with_seed(100, data.frame(
  Pop       = rep_len(utils::head(LETTERS, 10), 40),
  Sample    = sample(c("small", "medium", "large"), 40, replace = TRUE),
  TimePoint = rep(c("baseline", "6 months", "12 months", "24 months"), each = 10),
  seq.1234.56 = stats::rnorm(40, mean = 25, sd = 3.5)
))

df$TimePoint <- factor(df$TimePoint, levels = c("baseline", "6 months",
                                                "12 months", "24 months"))

new <- df |>  # Map 'TimePoint' var (character) to a numeric
 dplyr::left_join(data.frame(TimePoint = c("baseline", "6 months", "12 months", "24 months"),
                             time      = c(0,           6,          12,          24)
                             ),
                  by = "TimePoint") |>
 dplyr::mutate(id     = rep(1:10, times = 4),
               Sample = dplyr::case_when(id <= 4 ~ "small", # Recode Sample to correlate with IDs
                                         id > 7  ~ "large",
                                         TRUE    ~ "medium"))

# No title; no x-axis label; nothing fancy
plotLongitudinal(new, y = "seq.1234.56", time = "TimePoint", id = "id")

# Color lines by the 'Sample' variable
plotLongitudinal(new, y = "seq.1234.56", time = "TimePoint", id = "id",
                 color.by = "Sample")

# Color lines by the 'Sample' variable, using the 'time' numeric x variable
plotLongitudinal(new, y = "seq.1234.56", time = "time", id = "id",
                 color = "Sample")

# Use unquoted variable names
plotLongitudinal(new, y = seq.1234.56, time = time, id = id,
                 color = Sample)

# Use `mean` as the summary lines
plotLongitudinal(new, y = "seq.1234.56", id = "id", time = "time",
                 color = "Sample", summary.line = mean)

# Suppress summary line with `NULL`
plotLongitudinal(new, y = "seq.1234.56", id = "id", time = "time",
                 color.by = "Sample", summary.line = NULL,
                 main = "No Summary Lines")

# Boxplots by time point; `time` variable (TimePoint) must be a factor
plotLongitudinal(new, "seq.1234.56", id = "id", time = "TimePoint",
                 color.by = "Sample", add.box = TRUE, summary.line = NULL,
                 main = "With Time Point Boxplots")

# Facet by `Sample`
p <- plotLongitudinal(new, y = "seq.1234.56", id = "id", time = "time",
                      color.by = "Sample", summary.line = mean)

p + ggplot2::facet_wrap(~Sample) +
  ggplot2::ggtitle("Facet by `Sample`")

}
\seealso{
\code{\link[=geom_line]{geom_line()}}, \code{\link[=theme]{theme()}}
}
\author{
Stu Field
}
