% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcano.R
\name{plotVolcano}
\alias{plotVolcano}
\title{Create Volcano Plot}
\usage{
plotVolcano(
  data,
  FC,
  p.value,
  labels,
  identify = FALSE,
  fc.cutoff = 1,
  pt.size = 2.5,
  text.size = 3,
  cutoff = 0.05/nrow(data),
  main = NULL,
  x.lab = NULL,
  ...
)
}
\arguments{
\item{data}{A \code{data.frame} object containing at \emph{least} two columns
containing 1) log2-transformed fold-changes, 2) linear-space p-values.
A third column containing point labels \emph{must} be supplied,
if \code{identify = TRUE}.}

\item{FC}{Unquoted string identifying the column in \code{data} containing
a fold-change vector.}

\item{p.value}{Unquoted string identifying the column in \code{data} containing
a p-value vector.}

\item{labels}{Unquoted string identifying the column in \code{data} containing
point labels, typically \code{"Aptamer"} names.}

\item{identify}{Logical. Should significant points be identified?}

\item{fc.cutoff}{Placement for the cutoff for coloring points along the
fold-change x-axis. Defaults to doubling in fold-change (\code{1}).}

\item{pt.size}{Numeric. The size for the points.}

\item{text.size}{Numeric. The size for the identifying text.}

\item{cutoff}{Horizontal statistical significance cutoff for coloring
points. Defaults to Bonferroni corrected significance at \code{alpha = 0.05}
in "p-value" linear space \verb{[0, 1]}.}

\item{main}{Character. Main title for the plot.
See \code{\link[=ggtitle]{ggtitle()}} for \code{ggplot2} style graphics.}

\item{x.lab}{Character. Optional string for the x-axis. Otherwise
one is automatically generated (default).}

\item{...}{Arguments passed to \code{\link[=geom_point]{geom_point()}}.}
}
\description{
Create a volcano plot given a vector of log2-transformed
fold-changes (\code{FC}) and linear space p-values.
}
\examples{
withr::with_seed(101, {
  fc1 <- sort(runif(500, -2.5, 0))   # Z-scores as dummy fold-changes
  fc2 <- sort(runif(500, 0, 2.5))    # Z-scores as dummy fold-changes
  p1  <- pnorm(fc1)                  # p-values for neg. scores
  p2  <- pnorm(fc2, lower.tail = FALSE) # p-values for pos. scores
  p   <- jitter(c(p1, p2), amount = 0.1) # add noise
  p[p < 0] <- runif(sum(p < 0), 1e-05, 1e-02) # floor p > 0 after jitter
  df <- data.frame(fc = c(fc1, fc2), p = p)
})

plotVolcano(df, fc, p, cutoff = 0.1)  # lower p-value cutoff

# add some random labels to `df`
df <- dplyr::mutate(df, pt_label = dplyr::row_number())
plotVolcano(df, fc, p, labels = pt_label, identify = TRUE, cutoff = 0.01)
}
\seealso{
\code{\link[=geom_point]{geom_point()}}.
}
\author{
Stu Field
}
