% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcanoHTML.R
\name{plotVolcanoHTML}
\alias{plotVolcanoHTML}
\alias{plotVolcanoHTML.data.frame}
\alias{plotVolcanoHTML.stat_table}
\title{Generate HTML-based Volcano Plot}
\usage{
plotVolcanoHTML(data, FC, p.value, cutoff, fc.cutoff, main, x.lab, ...)

\method{plotVolcanoHTML}{data.frame}(
  data,
  FC,
  p.value,
  cutoff = 0.05/nrow(data),
  fc.cutoff = 1,
  main = NULL,
  x.lab = NULL,
  labels,
  ...
)

\method{plotVolcanoHTML}{stat_table}(
  data,
  FC,
  p.value,
  cutoff = 0.05/nrow(data$stat.table),
  fc.cutoff = 1,
  main = NULL,
  x.lab = NULL,
  tbl,
  ...
)
}
\arguments{
\item{data}{A data frame containing log2-transformed fold-changes and
corresponding p-values. An optional third column
containing, e.g. "Target" Names, can be passed if specified by
the \verb{labels =} parameter.}

\item{FC}{An unquoted string identifying the column in \code{data} containing
a vector \code{\link[=log2]{log2()}}-transformed fold-changes.}

\item{p.value}{An unquoted string identifying the column in \code{data} containing
a vector of p-values.}

\item{cutoff}{Horizontal statistical significance cutoff for coloring
points. Defaults to Bonferroni corrected significance at \code{alpha = 0.05}
in "p-value" linear space \verb{[0, 1]}.}

\item{fc.cutoff}{Placement for the cutoff for coloring points along the
fold-change x-axis. Defaults to doubling in fold-change (\code{1}).}

\item{main}{Character. Main title for the plot.
See \code{\link[=ggtitle]{ggtitle()}} for \code{ggplot2} style graphics.}

\item{x.lab}{Character. Optional string for the x-axis. Otherwise
one is automatically generated (default).}

\item{...}{Additional arguments passed to \code{\link[plotly:plotly]{plotly::plotly()}}.}

\item{labels}{An unquoted string identifying the column in \code{data} containing
point labels, typically "Target" or "Analyte" names.}

\item{tbl}{A \code{tibble} object containing analyte target annotation
information. This is usually the result of a call to \code{\link[SomaDataIO:getAnalyteInfo]{getAnalyteInfo()}}.}
}
\description{
Generate HTML-based Volcano Plot
}
\section{Methods (by class)}{
\itemize{
\item \code{plotVolcanoHTML(data.frame)}: Plot method for objects of class \code{data.frame}.

\item \code{plotVolcanoHTML(stat_table)}: Plot method for objects of class \code{stat_table} (SomaLogic internal).

}}
\examples{
# Dummy up a fake table with minimal variables
adat <- SomaDataIO::example_data
seqs <- SomaDataIO::getAnalytes(adat)
target_map <- SomaDataIO::getTargetNames(SomaDataIO::getAnalyteInfo(adat))
df <- withr::with_seed(101, {
  fc1 <- sort(runif(500, -2.5, 0))   # Z-scores as dummy fold-changes
  fc2 <- sort(runif(500, 0, 2.5))    # Z-scores as dummy fold-changes
  p1  <- pnorm(fc1)                  # p-values for neg. scores
  p2  <- pnorm(fc2, lower.tail = FALSE) # p-values for pos. scores
  p   <- jitter(c(p1, p2), amount = 0.1) # add noise
  p[p < 0] <- runif(sum(p < 0), 1e-05, 1e-02) # floor p > 0 after jitter
  seq_vec <- sample(seqs, length(p))   # random seqIds
  data.frame(
    AptName = seq_vec,
    t_stat  = runif(50, 10, 20),
    log2_fc = c(fc1, fc2),
    p_value = p,
    target  = unlist(target_map)[seq_vec]  # map random target names
  )
})

# S3 `data.frame` method
# No TargetNames -> `NA`
plotVolcanoHTML(df, log2_fc, p_value, cutoff = 0.1, fc.cutoff = 0.5)

# Add TargetNames via `labels=`
plotVolcanoHTML(df, log2_fc, p_value, cutoff = 0.1, fc.cutoff = 0.5, labels = target)
}
\seealso{
\code{\link[plotly:plotly]{plotly::plotly()}}

Other volcano: 
\code{\link{plotVolcano}()}
}
\author{
Leigh Alexander, Stu Field
}
\concept{volcano}
