% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z-deprecated.R
\name{SomaPlotr-deprecated}
\alias{SomaPlotr-deprecated}
\alias{calcOutlierMap}
\title{Deprecated function(s) of the \pkg{SomaPlotr} package}
\description{
These functions have either been
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
in the current version of \pkg{SomaPlotr} package.
Please re-code your scripts accordingly based on the
suggestions below:

\tabular{lcr}{
\strong{Function}       \tab                                    \tab \strong{Now Use} \cr
\code{\link[=calcOutlierMap]{calcOutlierMap()}}    \tab \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \tab \code{\link[SomaDataIO:calcOutlierMap]{SomaDataIO::calcOutlierMap()}} \cr
}
}
\details{
Some badges you may see in \pkg{SomaPlotr}:

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#soft-deprecated}{\figure{lifecycle-soft-deprecated.svg}{options: alt='[Soft-deprecated]'}}}{\strong{[Soft-deprecated]}}

\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
