% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SomaPlotr-package.R
\docType{package}
\name{SomaPlotr-package}
\alias{SomaPlotr}
\alias{SomaPlotr-package}
\title{SomaPlotr: Bioinformatics Graphics and Visualization of 'SomaScan' Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A highly specialized suite of standardized plotting routines based on the "Grammar of Graphics" framework of mapping variables to aesthetics used in 'ggplot2'. Graphics types are biased towards visualizing SomaScan (proteomic) data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://somalogic.github.io/SomaPlotr/}
  \item \url{https://somalogic.com}
  \item \url{https://ggplot2.tidyverse.org/}
  \item Report bugs at \url{https://github.com/SomaLogic/SomaPlotr/issues}
}

}
\author{
\strong{Maintainer}: Amanda Hiser \email{amanda.hiser@standardbio.com}

Other contributors:
\itemize{
  \item Standard BioTools, Inc. [copyright holder, funder]
}

}
\keyword{internal}
\keyword{package}
