% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-greys.R
\name{style-greys}
\alias{style-greys}
\alias{palette_soma_greys}
\alias{scale_color_soma_greys}
\alias{scale_fill_soma_greys}
\title{SomaLogic Greyscale Color Palette}
\usage{
palette_soma_greys(n)

scale_color_soma_greys(...)

scale_fill_soma_greys(...)
}
\arguments{
\item{n}{number of colors. This palette is set to the
\link{soma_colors_greys} object, which is currently 6 elements
in length.
The palette is recycled as necessary.}

\item{...}{Arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}},
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradient()}}, or \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradient()}}.}
}
\description{
The color palette of official SomaLogic greyscale colors.
}
\section{Functions}{
\itemize{
\item \code{scale_color_soma_greys()}: color \emph{discrete} scale for greyscale graphics.

\item \code{scale_fill_soma_greys()}: fill \emph{discrete} scale for greyscale graphics.

}}
\seealso{
Other palettes: 
\code{\link{style-gender}},
\code{\link{style-risk}},
\code{\link{style-soma}}
}
\concept{palettes}
