% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style-risk.R
\name{style-risk}
\alias{style-risk}
\alias{palette_soma_risk}
\alias{scale_color_soma_risk}
\alias{scale_fill_soma_risk}
\alias{scale_continuous_color_soma_risk}
\alias{scale_continuous_fill_soma_risk}
\title{SomaLogic Risk Category Color Palette}
\usage{
palette_soma_risk(direction = 4, n)

scale_color_soma_risk(direction = 4, ...)

scale_fill_soma_risk(direction = 4, ...)

scale_continuous_color_soma_risk(
  low = soma_colors_risk$green,
  high = soma_colors_risk$red,
  ...
)

scale_continuous_fill_soma_risk(
  low = soma_colors_risk$green,
  high = soma_colors_risk$red,
  ...
)
}
\arguments{
\item{direction}{Captures both number of colors and
direction of those colors based on \link{soma_colors_risk} colors,
which are currently 5 colors from high to low risk.
The palette is recycled as necessary.
A positive number maintains order from high (red) to
low (green) risk. A negative number reverses that order.
There is special handling for \verb{n = 2, 3, and 4} colors for
both directions. For more custom ordering, use
\code{\link[=scale_color_manual]{scale_color_manual()}} or \code{\link[=scale_fill_manual]{scale_fill_manual()}} with
the \link{soma_colors_risk} colors.}

\item{n}{Not currently used. Included to make palette function work
with \code{\link[=scale_color_soma_risk]{scale_color_soma_risk()}} and \code{\link[=scale_fill_soma_risk]{scale_fill_soma_risk()}}.}

\item{...}{Arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}},
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradient()}}, or \code{\link[ggplot2:scale_gradient]{ggplot2::scale_color_gradient()}}.}

\item{low, high}{Colours for low and high ends of the gradient.}
}
\description{
The color palette of official SomaLogic risk categories/trends.
}
\section{Functions}{
\itemize{
\item \code{scale_color_soma_risk()}: color \emph{discrete} scale for risk categories.

\item \code{scale_fill_soma_risk()}: fill \emph{discrete} scale for risk categories.

\item \code{scale_continuous_color_soma_risk()}: color \emph{continuous} scale for risk trends.
Currently implemented for 2 colors (high risk red to low
risk green). Multiple colors can be done in a gradient with
\code{\link[=scale_color_gradientn]{scale_color_gradientn()}}.

\item \code{scale_continuous_fill_soma_risk()}: fill \emph{continuous} scale for risk trends.

}}
\seealso{
Other palettes: 
\code{\link{style-gender}},
\code{\link{style-greys}},
\code{\link{style-soma}}
}
\concept{palettes}
