% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotBeeswarm.R
\name{boxplotBeeswarm}
\alias{boxplotBeeswarm}
\title{Create a "Beeswarm" Boxplot}
\usage{
boxplotBeeswarm(
  .data,
  notch = FALSE,
  label = "Group",
  main = NULL,
  y.lab = "value",
  x.lab = label,
  cols,
  pt.size = 2.5,
  pt.color = "black",
  pt.shape = 21,
  ...
)
}
\arguments{
\item{.data}{Either a \code{data.frame}/\code{tbl_df} object where each column is
a numeric vector containing values for each box, or a \emph{named}
\code{list} object which can be converted to one.}

\item{notch}{Logical. Should notches be drawn in the boxplots? If FALSE
(default), a standard box plot will be drawn. If TRUE, notches will
be added to the median line of each box plot. See \code{notch} argument of
\code{\link[=geom_boxplot]{geom_boxplot()}}.}

\item{label}{Character. A label for the grouping variable, i.e. what
the columns of the data frame represent.}

\item{main}{Character. Main title for the plot.
See \code{\link[=ggtitle]{ggtitle()}} for \code{ggplot2} style graphics.}

\item{y.lab}{Character. Optional string for the y-axis. Otherwise
one is automatically generated (default).}

\item{x.lab}{Character. Optional string for the x-axis. Otherwise
one is automatically generated (default).}

\item{cols}{Character. A \emph{vector} of colors for the groups/boxes.
For \code{\link[=plotDoubleHist]{plotDoubleHist()}}, must be \code{length = 2}.}

\item{pt.size}{Numeric. A size for the points. See \code{\link[=geom_point]{geom_point()}}.}

\item{pt.color}{Character. A \emph{fill} color for the points. See \code{\link[=geom_point]{geom_point()}}.}

\item{pt.shape}{Numeric or Character. Recognized \code{pch} shapes for the
points. Recall that \emph{only} \code{pch = 21 - 25} are "fill-able". Other point
characters will \emph{not} take on the color from \code{pt.color}.
See \code{\link[=geom_point]{geom_point()}}.}

\item{...}{Additional arguments passed to \code{\link[=geom_boxplot]{geom_boxplot()}}.}
}
\value{
Boxplot with "beeswarm" style points.
}
\description{
Plot a series of boxplots with "beeswarm"-style points added to the boxes.
}
\examples{
df <- lapply(setNames(LETTERS[1:5], letters[1:5]), \(x) rnorm(30, 10, 3)) |>
  data.frame() |>
  dplyr::select(d, dplyr::everything())   # move `d` to the front
head(df)

df |> boxplotBeeswarm(main = "Title")
df |> boxplotBeeswarm(pt.color = "cyan", notch = TRUE) # add notch
df |> boxplotBeeswarm(cols = "grey")     # all boxes 1 color
df |> boxplotBeeswarm(label = "Disease Level", y.lab = "Y")

# Some point shapes can't be filled with color.
# Example of a non-fillable `pt.shape` value
df |> boxplotBeeswarm(pt.shape = 13)

# Shapes 21 - 25 are fillable
df |> boxplotBeeswarm(cols = rep("blue", ncol(df)), pt.size = 5,
                       pt.shape = 23, pt.color = "red")
}
\seealso{
\code{\link[=geom_boxplot]{geom_boxplot()}}, \code{\link[=geom_jitter]{geom_jitter()}}

Other boxplots: 
\code{\link{boxplotGrouped}()},
\code{\link{boxplotSubarray}()}
}
\author{
Stu Field
}
\concept{boxplots}
