% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotSubarray.R
\name{boxplotSubarray}
\alias{boxplotSubarray}
\title{Plot Boxplots by Subarray (Sample)}
\usage{
boxplotSubarray(
  .data,
  color.by = NULL,
  labels = "SampleId",
  y.lim = NULL,
  do.log = TRUE,
  apts = NULL
)
}
\arguments{
\item{.data}{A \code{soma_data} or data frame object, created from a SomaScan
ADAT file, via a call to \code{\link[=read_adat]{read_adat()}}. This object must contain the
following columns: \code{PlateId}, \code{SampleId}, \code{SampleType}, \code{SampleMatrix},
\code{Barcode2d}, \code{SlideId}, \code{Subarray}, and \code{HybControlNormScale}.}

\item{color.by}{Character. A column name to color the subarrays (samples)
by. This is typically a sample processing or clinical data field in the
ADAT such as \code{SlideId}.}

\item{labels}{Character. The column name of \code{.data} used to label each box.}

\item{y.lim}{Numeric. Length 2. The upper- and lower-quantiles of the
\emph{total} data used to determine the y-axis limits of the plot.
If \code{NULL}, all points are shown.}

\item{do.log}{Logical. Should the data be log10-transformed?}

\item{apts}{Optional. A subset of analytes (as \code{AptNames}) to add as points
on top of the subarray boxplot.}
}
\description{
Plots the distribution of all analytes, stratified by subarray, as a
boxplot. These plots are intended to be used as a quality control
visualization tool for SomaScan assay runs. In SomaScan (\code{soma_adat}) data
format, the term "subarray" is analogous to sample, and typically indicates
a row in the data.
}
\examples{
data <- SomaDataIO::example_data

# Randomly select a small subset of samples
s_rn <- withr::with_seed(101, sample(rownames(data), 30L))
data <- data[s_rn, ]
boxplotSubarray(data)

# Color by `SampleType` variable
boxplotSubarray(data, color.by = "SampleType")

# Find the feature names of the corresponding hyb controls
hybs <- SomaDataIO::getAnalyteInfo(data) |>
  dplyr::filter(grepl("^Hybridization", Type)) |>
  dplyr::pull(AptName)

# Pass hyb controls to `apts` arg; plots the points in boxes/whiskers
boxplotSubarray(data, color.by = "SampleType", apts = hybs)

# Zoom to (20, 80) quantiles
boxplotSubarray(data, color.by = "SampleType", y.lim = c(0.2, 0.8))

# Group by an additional variable
boxplotSubarray(data, color.by = "SampleType") +
  ggplot2::facet_wrap(~PlateId)
}
\seealso{
\code{\link[=geom_boxplot]{geom_boxplot()}}

Other boxplots: 
\code{\link{boxplotBeeswarm}()},
\code{\link{boxplotGrouped}()}
}
\author{
Stu Field
}
\concept{boxplots}
