% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotVolcanoHTML.R
\name{plotVolcanoHTML}
\alias{plotVolcanoHTML}
\alias{plotVolcanoHTML.data.frame}
\alias{plotVolcanoHTML.stat_table}
\title{Generate HTML-based Volcano Plot}
\usage{
plotVolcanoHTML(data, FC, p.vec, cutoff, fc.cutoff, main, x.lab, ...)

\method{plotVolcanoHTML}{data.frame}(
  data,
  FC = signed.log2.fold.change,
  p.vec = p.value,
  cutoff = 0.05/nrow(data),
  fc.cutoff = 1,
  main = NULL,
  x.lab = NULL,
  target.labels,
  ...
)

\method{plotVolcanoHTML}{stat_table}(
  data,
  FC = signed.log2.fold.change,
  p.vec = p.value,
  cutoff = 0.05/nrow(data$stat.table),
  fc.cutoff = 1,
  main = NULL,
  x.lab = NULL,
  apt.data,
  ...
)
}
\arguments{
\item{data}{One of two object classes:
\describe{
\item{\code{data.frame}:}{Containing log2-transformed fold-changes and
corresponding p-values. In this case, an optional third column
containing Target Names can be passed if specified in
the \verb{target.labels =} argument.}
\item{\code{stat_table}:}{An object of class \verb{c(stat_table, *_table)}.
When this argument is provided, the \code{apt.data} argument \emph{must}
also be passed.}
}}

\item{FC}{An un-quoted variable name in \code{data} for the fold-changes.}

\item{p.vec}{An un-quoted variable name in \code{data} for the p-values.}

\item{cutoff}{Horizontal cutoff for coloring points. Defaults to Bonferroni
corrected level of significance.}

\item{fc.cutoff}{Vertical cutoff for coloring coloring points
(i.e. the fold-change). Defaults to doubling in fold-change.}

\item{main}{Character. Main title for the plot.
See \code{\link[=ggtitle]{ggtitle()}} for \code{ggplot2} style graphics.}

\item{x.lab}{Character. Optional string for the x-axis. Otherwise
one is automatically generated (default).}

\item{...}{Additional arguments passed to \code{\link[plotly:plotly]{plotly::plotly()}}.}

\item{target.labels}{Optional for the \code{data.frame} method. An un-quoted
variable name in \code{data} for the Target Names.}

\item{apt.data}{An \code{apt.data} object containing target information
used to generate point labels in the final plot.
This object can be obtained from a call to \code{\link[=getAnalyteInfo]{getAnalyteInfo()}}
on the original data.}
}
\description{
Generate HTML-based Volcano Plot
}
\section{Methods (by class)}{
\itemize{
\item \code{plotVolcanoHTML(data.frame)}: Plot method for objects of class \code{data.frame}.

\item \code{plotVolcanoHTML(stat_table)}: Plot method for objects of class \code{stat_table}.

}}
\examples{
# Dummy up a fake table with minimal variables
data <- SomaDataIO::example_data
apts <- SomaDataIO::getAnalytes(data)
df <- withr::with_seed(100,
   data.frame(
              AptName = sample(apts, 50),
              t.stat = runif(50, 10, 20),
              p.value = runif(50),
              signed.log2.fold.change = rnorm(50),
              Zscore = rnorm(50),
              # generate fake target names
              target = paste0("Target", "_", sample(LETTERS, 50, replace = TRUE),
                              sample(1:50, 50), "-",
                              sample(letters, 50, replace = TRUE))
             )
)

# S3 `data.frame` method
# No Target labels -> `NA`
plotVolcanoHTML(df, cutoff = 0.1, fc.cutoff = 0.5)

# Add Target labels -> `targets`
plotVolcanoHTML(df, cutoff = 0.1, fc.cutoff = 0.5, target.labels = target)

# S3 `stat_table` method; calc.* family
# Dummy up a fake `stat_table` object
stbl <- list()
stbl$stat.table <- SomaDataIO::col2rn(df, "AptName")
stbl$test <- "Fake t-test"
stbl$data.dim <- dim(df)
stbl$y.response <- "LHS"
stbl$counts <- c(GroupA = 25, GroupB = 25)
stbl$log <- TRUE
stbl$data.frame <- "df"
class(stbl) <- c("stat_table", class(stbl))

# Pick up apt.data
ad <- SomaDataIO::getAnalyteInfo(data)

# Target names picked up from `apt.data`
plotVolcanoHTML(stbl, apt.data = ad, cutoff = 0.1, fc.cutoff = 0.5)

# Change FC arg on fly -> `Zscore`
# Pass args via '...' -> `x.lab`
plotVolcanoHTML(stbl, apt.data = ad, cutoff = 0.1, fc.cutoff = 0.5,
                FC = Zscore, x.lab = "Slope")
}
\seealso{
\code{\link[plotly:plotly]{plotly::plotly()}}
}
\author{
Leigh Alexander, Stu Field
}
